product-type-id working but not worth it <>
- we get 15% smaller files or mellite binary export
- and if applying zip, the gain is almost zero

+++ api/shared/src/main/scala/de/sciss/synth/ugen/BasicOpUGen.scala
@@ -191,7 +191,7 @@ object UnaryOpUGen extends ProductType[UnaryOpUGen] {
   // case object NotNil      extends Op(  3 )
   case object BitNot extends PureOp {
     final val id = 4
-    def make1(a: Float): Float = ~a.toInt
+    def make1(a: Float): Float = (~a.toInt).toFloat
   }
   case object Abs extends PureOp {
     final val id = 5
